// src/generated/models/warehouse-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BFF Service
 * Backend For Frontend service aggregating microservices
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: you@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface WarehouseResponse
 */
export interface WarehouseResponse {
    /**
     * 
     * @type {string}
     * @memberof WarehouseResponse
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof WarehouseResponse
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof WarehouseResponse
     */
    'country': string;
    /**
     * 
     * @type {string}
     * @memberof WarehouseResponse
     */
    'city': string;
    /**
     * 
     * @type {string}
     * @memberof WarehouseResponse
     */
    'address': string;
    /**
     * 
     * @type {string}
     * @memberof WarehouseResponse
     */
    'created_at': string;
    /**
     * 
     * @type {string}
     * @memberof WarehouseResponse
     */
    'updated_at': string;
}